requirements.txt
flask~=3.1.2
flasgger~=0.9.7.1
sqlalchemy~=2.0.43
openai~=1.97.1
python-dotenv~=1.0.1
flask-migrate~=4.1.0
flask-cors~=6.0.1
alembic~=1.16.5